emotion_classify/requirements.txt
torch
fastapi
sqlalchemy
transformers
pydantic
streamlit
requests
matplotlib.pyplot
uvicorn